fix: Corrige criacao de usuario

+++ frontend/src/pages/PageCadastroUsuario.vue
@@ -1,8 +1,11 @@
 <template>
-  <div class="flex flex-col p-24">
-    <q-form class="flex flex-col gap-16">
+  <div class="flex flex-col p-24  items-center justify-center gap-16 max-w-[40%] sm:max-w-full m-auto h-full overflow-hidden">
+    <img :src="logo" alt="Logo simply+" class="w-[200px] h-[200px]" />
+
+    <q-form class="flex flex-col gap-16 w-full">
       <q-input
         filled
+        class="w-full"
         v-model="model.email"
         label="Email *"
         lazy-rules
@@ -11,17 +14,18 @@
 
       <q-input
         filled
+        class="w-full"
         v-model="model.nome"
         label="Nome *"
         lazy-rules
         :rules="[(val) => (val && val.length > 0) || 'Campo obrigatório']"
       />
 
-      <q-input filled type="password" v-model="model.senha" label="Senha *" lazy-rules :rules="[(val) => (val && val.length > 0) || 'Campo obrigatório']"/>
+      <q-input class="w-full" filled type="password" v-model="model.senha" label="Senha *" lazy-rules :rules="[(val) => (val && val.length > 0) || 'Campo obrigatório']"/>
 
       <div class="flex w-full gap-20 justify-between items-center">
-        <q-btn label="Limpar" type="button" @click="reset()" color="secondary" flat />
-        <q-btn label="Cadastrar" type="button" @click="criar()" color="primary" />
+        <q-btn label="Limpar" type="button" @click="reset()" class="text-secondary" flat />
+        <q-btn label="Cadastrar" type="button" @click="criar()" class="bg-primary-pure text-[#fff]" />
       </div>
     </q-form>
   </div>
@@ -30,6 +34,7 @@
 <script setup>
 import { ref } from 'vue'
 import useUsuario from '../store/usuario'
+import logo from '../../public/logo.png'
 
 const { addNewItemUsuarios } = useUsuario()
 
@@ -37,6 +42,8 @@ const model = ref({ email: '', senha: '', nome: '', data_criacao: '' })
 
 const criar = () => {
   const usuariosList = JSON.parse(localStorage.getItem('usuarios')) || []
+  const userExiste = usuariosList.find(user => user.senha === model.value.senha && user.email === model.value.email)
+  if(userExiste) return
   localStorage.setItem(
     'usuarios',
     JSON.stringify([
